+++ src/components/main/main-styled.tsx
@@ -1,14 +1,4 @@
-import { ReactNode } from 'react'
 import styled, { css } from 'styled-components/macro'
-
-type MainProps = {
-  children: ReactNode | ReactNode[]
-}
-
-function Main({ children }: MainProps) {
-  return <MainStyled>{children}</MainStyled>
-}
-
 const MainStyled = styled.main`
   display: grid;
   grid-template-columns: 332px auto;
@@ -19,5 +9,4 @@ const MainStyled = styled.main`
     background: ${theme.colors.white};
   `}
 `
-
-export default Main
+export default MainStyled
