+++ src/components/sidebar/sidebar.tsx
@@ -1,0 +1,12 @@
+import { Logo } from 'components/logo'
+import SidebarStyled from './sidebar-styled'
+
+function Sidebar() {
+  return (
+    <SidebarStyled>
+      <Logo />
+    </SidebarStyled>
+  )
+}
+
+export { Sidebar }
